requirements.txt
pdfminer.six pdfminer
openpyxl openpyxl
pandas pandas
